Movement in x and y

+++ PROItheGame/PROItheGame/RectangularObject.h
@@ -6,23 +6,24 @@ class RectangularObject :
 	public AbstractObject
 {
 protected:
-	int xCoordinate;
-	int yCoordinate;
+	double xCoordinate;
+	double yCoordinate;
 	int objectHeight;
 	int objectWidth;
+
 public:
 	RectangularObject(int x = 0, int y = 0, int h = 0, int w = 0) : xCoordinate(x), yCoordinate(y), objectHeight(h), objectWidth(w) {};
 	~RectangularObject();
 
-	//setters
-	void setXCoordinate(int newValue);
-	void setYCoordinate(int newValue);
+	///setters
+	void setXCoordinate(double newValue);
+	void setYCoordinate(double newValue);
 	void setObjectHeight(int newValue);
 	void setObjectWidth(int newValue);
 
-	//getters
-	int getXCoordinate();
-	int getYCoordinate();
+	///getters
+	double getXCoordinate();
+	double getYCoordinate();
 	int getObjectHeight();
 	int getObjectWidth();
